feat: V2: add multi thread and async functions from tokio
